Move markdown-it parsing to /markbind
Markdown passed in attributes are being rendered in
the client.

This results in a dependency on the markdown parsing
library ( markdown-it ) on the client bloating the resulting
bundle size for /vue-strap. It also slows webpage loading,
especially on older devices. Moreover, /markbind already
parses some amount of markdown in attributes for other
purposes when building the site. This results in a less
cohesive codebase.

Let's rewrite the templates to use vue's slots for
displaying the rendered html previously present
in attributes.