Organize searchbar interface results by page
Search results belonging to the same page display as individual entries,
showing the page title multiple times.
Results are also not organized by page, but by the total number of
matches in the page title, heading and keywords with the search term.
When a page title is not specified in the site config, the search result
displays an empty block of vertical space.

This can lead to a less pleasant user experience with the searchbar.

Let’s redesign the searchbar user interface, organizing results by page.
Results belonging to the same page are sorted by the number of
matches, and pages are sorted according to the total number of its
matches in the page’s headings and keywords.
Let’s use the page’s src as a fallback title if the title is absent.