Add panel expand and collapse transitions
When a panel is expanded or collapsed, the panel’s content is also
instantly shown or hidden.

Moreover, when a panel is collapsed, the scroll position of the browser
stays the same, possibly showing the user a distant area of the page.
This can cause the reader to feel lost after collapsing the panel.

Let’s add panel expand and collapse transitions to panels.
Additionally, when a panel is collapsed, the reader’s scroll position is
smoothly animated back to the collapsed panel if the top of the panel
is out of the reader’s view.