Add versions component for versioning feature
To allow changes versioning user interface across versions easily, the
versions previously archived by the user are stored in versions.json.

Hence, to encapsulate the fetching of this json file and the
corresponding user interface of it, let’s implement a simple component
that takes after the style of markbind’s sitenav.
Let’s also provide a dropdown styled component so that it can be used
in the navigation bar instead if desired.